Loosen OpenTelemetry dependency version constraint (#216)
* Loosen OpenTelemetry dependency version constraint

When using the datacontract-cli as a library dependency, OpenTelemetry
may well be already a dependency of the dependee. The previous
dependency constraint allowed only other patch versions of OpenTelemetry
to be used. However, other minor versions should also be compatible.

As such, the dependency constraints for the OpenTelemetry OTLP exporters
have been loosened to allow minor version upgrades.

* Add setuptools dependency

Soda requires the `distutils` module. As this has been removed with
Python 3.12, `setuptools` is required as a drop-in replacement.

* Update pyproject.toml

---------

Co-authored-by: jochenchrist <jochenchrist@gmail.com>